+++ example/producer_avro/main.go
@@ -10,8 +10,8 @@ import (
 	"github.com/zillow/zkafka"
 )
 
-//go:embed dummy_event.avsc
-var dummyEventSchema string
+//go:embed event.avsc
+var eventSchema string
 
 func main() {
 	ctx := context.Background()
@@ -29,9 +29,7 @@ func main() {
 				AutoRegisterSchemas: true,
 				// When using avro schema registry, you must specify the schema. In this case,
 				// the schema used to generate the golang type is used.
-				// The heetch generated struct also embeds the schema as well (and isn't lossy like some of the
-				// other generative solutions. For example, one lib didn't include default values), so that could be used as well.
-				Schema: dummyEventSchema,
+				Schema: eventSchema,
 			},
 		},
 	})
@@ -40,6 +38,9 @@ func main() {
 		log.Panic(err)
 	}
 	for {
+		// The DummyEvent type is generated using `hamba/avro` (see make). This is the preferred generation for
+		// `formatter=zkafka.AvroSchemaRegistry` because the underlying serializer uses the avro tags on the generated struct
+		//	// to properly connect the schema and struct
 		event := DummyEvent{
 			IntField:    rand.Intn(100),
 			StringField: randomNames[rand.Intn(len(randomNames))],