+++ test/schema_registry_evo_test.go
@@ -12,6 +12,7 @@ import (
 	"fmt"
 	"math/rand"
 	"testing"
+	"time"
 
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/google/uuid"
@@ -83,27 +84,24 @@ func Test_SchemaRegistryReal_Avro_AutoRegisterSchemas_BackwardCompatibleSchemasC
 		},
 	})
 	require.NoError(t, err)
+	id := uuid.NewString()
 
-	evt1 := avro1.DummyEvent{
-		IntField:    int(rand.Int31()),
-		DoubleField: rand.Float64(),
-		StringField: uuid.NewString(),
-		BoolField:   true,
-		BytesField:  []byte(uuid.NewString()),
+	evt1 := avro1.Event{
+		ID:                     id,
+		DeliveredAtDateTimeUtc: time.Now().UTC().Truncate(time.Millisecond),
+		EventType:              "created",
 	}
-	// write msg1, and msg2
-	_, err = writer1.Write(ctx, evt1)
+	_, err = writer1.Write(ctx, &evt1)
 	require.NoError(t, err)
 
-	evt2 := avro2.DummyEvent{
-		IntField:            int(rand.Int31()),
-		DoubleField:         rand.Float64(),
-		StringField:         uuid.NewString(),
-		BoolField:           true,
-		BytesField:          []byte(uuid.NewString()),
-		NewFieldWithDefault: ptr(uuid.NewString()),
+	listingID2 := uuid.NewString()
+
+	evt2 := avro2.Event{
+		ID:                     listingID2,
+		DeliveredAtDateTimeUtc: time.Now().UTC().Truncate(time.Millisecond),
+		EventType:              "created",
 	}
-	_, err = writer2.Write(ctx, evt2)
+	_, err = writer2.Write(ctx, &evt2)
 	require.NoError(t, err)
 
 	consumerTopicConfig := zkafka.ConsumerTopicConfig{
@@ -111,7 +109,8 @@ func Test_SchemaRegistryReal_Avro_AutoRegisterSchemas_BackwardCompatibleSchemasC
 		Topic:     topic,
 		Formatter: zkafka.AvroSchemaRegistry,
 		SchemaRegistry: zkafka.SchemaRegistryConfig{
-			URL: "http://localhost:8081",
+			URL:             "http://localhost:8081",
+			Deserialization: zkafka.DeserializationConfig{Schema: dummyEventSchema1},
 		},
 		GroupID: groupID,
 		AdditionalProps: map[string]any{
@@ -131,22 +130,20 @@ func Test_SchemaRegistryReal_Avro_AutoRegisterSchemas_BackwardCompatibleSchemasC
 
 	require.NoError(t, reader.Close())
 
-	receivedEvt1 := avro1.DummyEvent{}
+	receivedEvt1 := avro1.Event{}
 	require.NoError(t, msg1.Decode(&receivedEvt1))
-	assertEqual(t, evt1, receivedEvt1)
+	assertEqual(t, receivedEvt1, evt1)
 
-	receivedEvt2Schema1 := avro1.DummyEvent{}
+	receivedEvt2Schema1 := avro1.Event{}
 	require.NoError(t, msg2.Decode(&receivedEvt2Schema1))
-	expectedEvt2 := avro1.DummyEvent{
-		IntField:    evt2.IntField,
-		DoubleField: evt2.DoubleField,
-		StringField: evt2.StringField,
-		BoolField:   evt2.BoolField,
-		BytesField:  evt2.BytesField,
+	expectedEvt2 := avro1.Event{
+		ID:                     evt2.ID,
+		DeliveredAtDateTimeUtc: evt2.DeliveredAtDateTimeUtc,
+		EventType:              evt2.EventType,
 	}
 	assertEqual(t, expectedEvt2, receivedEvt2Schema1)
 
-	receivedEvt2Schema2 := avro2.DummyEvent{}
+	receivedEvt2Schema2 := avro2.Event{}
 	require.NoError(t, msg2.Decode(&receivedEvt2Schema2))
 	assertEqual(t, evt2, receivedEvt2Schema2)
 }
@@ -369,117 +366,6 @@ func Test_SchemaRegistryReal_JSON_AutoRegisterSchemas_BackwardCompatibleSchemasC
 	receivedEvt2Schema2 := json2.DummyEvent{}
 	require.NoError(t, msg2.Decode(&receivedEvt2Schema2))
 	assertEqual(t, evt2, receivedEvt2Schema2, cmpopts.IgnoreUnexported(json2.DummyEvent{}))
-}
-
-func Test_SchemaRegistry_Avro_AutoRegisterSchemas_BackwardCompatibleSchemasCanBeRegisteredAndReadFrom(t *testing.T) {
-	checkShouldSkipTest(t, enableKafkaBrokerTest)
-
-	ctx := context.Background()
-	topic := "integration-test-topic-2" + uuid.NewString()
-	bootstrapServer := getBootstrap()
-
-	createTopic(t, bootstrapServer, topic, 1)
-	t.Logf("Created topic: %s", topic)
-
-	groupID := uuid.NewString()
-
-	client := zkafka.NewClient(zkafka.Config{BootstrapServers: []string{bootstrapServer}}, zkafka.LoggerOption(stdLogger{}))
-	defer func() { require.NoError(t, client.Close()) }()
-
-	t.Log("Created writer with auto registered schemas")
-	writer1, err := client.Writer(ctx, zkafka.ProducerTopicConfig{
-		ClientID:  fmt.Sprintf("writer-%s-%s", t.Name(), uuid.NewString()),
-		Topic:     topic,
-		Formatter: zkafka.AvroSchemaRegistry,
-		SchemaRegistry: zkafka.SchemaRegistryConfig{
-			URL: "mock://",
-			Serialization: zkafka.SerializationConfig{
-				AutoRegisterSchemas: true,
-				Schema:              dummyEventSchema1,
-			},
-		},
-	})
-	require.NoError(t, err)
-
-	writer2, err := client.Writer(ctx, zkafka.ProducerTopicConfig{
-		ClientID:  fmt.Sprintf("writer-%s-%s", t.Name(), uuid.NewString()),
-		Topic:     topic,
-		Formatter: zkafka.AvroSchemaRegistry,
-		SchemaRegistry: zkafka.SchemaRegistryConfig{
-			URL: "mock://",
-			Serialization: zkafka.SerializationConfig{
-				AutoRegisterSchemas: true,
-				Schema:              dummyEventSchema2,
-			},
-		},
-	})
-	require.NoError(t, err)
-
-	evt1 := avro1.DummyEvent{
-		IntField:    int(rand.Int31()),
-		DoubleField: rand.Float64(),
-		StringField: uuid.NewString(),
-		BoolField:   true,
-		BytesField:  []byte(uuid.NewString()),
-	}
-	// write msg1, and msg2
-	_, err = writer1.Write(ctx, evt1)
-	require.NoError(t, err)
-
-	evt2 := avro2.DummyEvent{
-		IntField:            int(rand.Int31()),
-		DoubleField:         rand.Float64(),
-		StringField:         uuid.NewString(),
-		BoolField:           true,
-		BytesField:          []byte(uuid.NewString()),
-		NewFieldWithDefault: ptr(uuid.NewString()),
-	}
-	_, err = writer2.Write(ctx, evt2)
-	require.NoError(t, err)
-
-	consumerTopicConfig := zkafka.ConsumerTopicConfig{
-		ClientID:  fmt.Sprintf("reader-%s-%s", t.Name(), uuid.NewString()),
-		Topic:     topic,
-		Formatter: zkafka.AvroSchemaRegistry,
-		SchemaRegistry: zkafka.SchemaRegistryConfig{
-			URL: "mock://",
-		},
-		GroupID: groupID,
-		AdditionalProps: map[string]any{
-			"auto.offset.reset": "earliest",
-		},
-	}
-	reader, err := client.Reader(ctx, consumerTopicConfig)
-	require.NoError(t, err)
-
-	t.Log("Begin reading messages")
-	results, err := readMessages(reader, 2)
-	require.NoError(t, err)
-
-	msg1 := <-results
-	msg2 := <-results
-	t.Log("Close reader")
-
-	require.NoError(t, reader.Close())
-
-	receivedEvt1 := avro1.DummyEvent{}
-	require.NoError(t, msg1.Decode(&receivedEvt1))
-	assertEqual(t, evt1, receivedEvt1)
-
-	receivedEvt2Schema1 := avro1.DummyEvent{}
-	require.NoError(t, msg2.Decode(&receivedEvt2Schema1))
-	expectedEvt2 := avro1.DummyEvent{
-		IntField:    evt2.IntField,
-		DoubleField: evt2.DoubleField,
-		StringField: evt2.StringField,
-		BoolField:   evt2.BoolField,
-		BytesField:  evt2.BytesField,
-	}
-	assertEqual(t, expectedEvt2, receivedEvt2Schema1)
-
-	receivedEvt2Schema2 := avro2.DummyEvent{}
-	require.NoError(t, msg2.Decode(&receivedEvt2Schema2))
-	assertEqual(t, evt2, receivedEvt2Schema2)
 }
 
 func Test_SchemaRegistry_Proto_AutoRegisterSchemas_BackwardCompatibleSchemasCanBeRegisteredAndReadFrom(t *testing.T) {